+++ python/GafferUITest/ContextTrackerTest.py
@@ -991,5 +991,114 @@ class ContextTrackerTest( GafferUITest.TestCase ) :
 		self.assertTrue( tracker.isEnabled( script["tracked" ] ) )
 		self.assertFalse( tracker.isEnabled( script["untracked" ] ) )
 
+	def testContextForEditorSettings( self ) :
+
+		script = Gaffer.ScriptNode()
+
+		script["node"] = GafferTest.AddNode()
+
+		script["contextVariables"] = Gaffer.ContextVariables()
+		script["contextVariables"].setup( script["node"]["sum"] )
+		script["contextVariables"]["in"].setInput( script["node"]["sum"] )
+		script["contextVariables"]["variables"].addChild( Gaffer.NameValuePlug( "testVariable", True ) )
+
+		script.setFocus( script["contextVariables"] )
+
+		editor = GafferUI.Editor( GafferUI.Label( "TestEditor" ), script )
+		editor.settings()["in"] = Gaffer.IntPlug()
+		editor.settings()["testPlug"] = Gaffer.StringPlug()
+
+		with self.UpdateHandler() :
+			tracker = GafferUI.ContextTracker.acquireForFocus( script )
+
+		self.assertIn( "testVariable", tracker.context( script["node"] ) )
+		self.assertNotIn( "testVariable", tracker.context( script["contextVariables"] ) )
+
+		# No node being viewed.
+
+		for plug in ( editor.settings()["in"], editor.settings()["testPlug"] ) :
+			self.assertFalse( tracker.isTracked( plug ) )
+			self.assertEqual( tracker.context( plug ), script.context() )
+
+		# Node being viewed. We want to take the context from that.
+
+		genericChangedSlot = GafferTest.CapturingSlot( tracker.changedSignal() )
+		specificChangedSlot = GafferTest.CapturingSlot( tracker.changedSignal( editor.settings()["testPlug" ] ) )
+
+		editor.settings()["in"].setInput( script["node"]["sum"] )
+
+		self.assertEqual( len( genericChangedSlot ), 0 )
+		self.assertEqual( len( specificChangedSlot ), 1 )
+
+		for plug in ( editor.settings()["in"], editor.settings()["testPlug"] ) :
+			self.assertFalse( tracker.isTracked( plug ) )
+			self.assertEqual( tracker.context( plug ), tracker.context( script["node"] ) )
+
+		# Edit the context. This should also emit the editor-specific signal.
+
+		with self.UpdateHandler() :
+			script["contextVariables"]["variables"][0]["value"].setValue( False )
+
+		self.assertEqual( len( genericChangedSlot ), 1 )
+		self.assertEqual( len( specificChangedSlot ), 2 )
+
+		for plug in ( editor.settings()["in"], editor.settings()["testPlug"] ) :
+			self.assertFalse( tracker.isTracked( plug ) )
+			self.assertEqual( tracker.context( plug ), tracker.context( script["node"] ) )
+
+	def testContextForView( self ) :
+
+		script = Gaffer.ScriptNode()
+
+		script["node"] = GafferTest.AddNode()
+
+		script["contextVariables"] = Gaffer.ContextVariables()
+		script["contextVariables"].setup( script["node"]["sum"] )
+		script["contextVariables"]["in"].setInput( script["node"]["sum"] )
+		script["contextVariables"]["variables"].addChild( Gaffer.NameValuePlug( "testVariable", True ) )
+
+		script.setFocus( script["contextVariables"] )
+
+		view = GafferUITest.ViewTest.MyView( script )
+		view["testPlug"] = Gaffer.StringPlug()
+
+		with self.UpdateHandler() :
+			tracker = GafferUI.ContextTracker.acquireForFocus( script )
+
+		self.assertIn( "testVariable", tracker.context( script["node"] ) )
+		self.assertNotIn( "testVariable", tracker.context( script["contextVariables"] ) )
+
+		# No node being viewed.
+
+		for plug in ( view["in"], view["testPlug"] ) :
+			self.assertFalse( tracker.isTracked( plug ) )
+			self.assertEqual( tracker.context( plug ), script.context() )
+
+		# Node being viewed. We want to take the context from that.
+
+		genericChangedSlot = GafferTest.CapturingSlot( tracker.changedSignal() )
+		specificChangedSlot = GafferTest.CapturingSlot( tracker.changedSignal( view["testPlug" ] ) )
+
+		view["in"].setInput( script["node"]["sum"] )
+
+		self.assertEqual( len( genericChangedSlot ), 0 )
+		self.assertEqual( len( specificChangedSlot ), 1 )
+
+		for plug in ( view["in"], view["testPlug"] ) :
+			self.assertFalse( tracker.isTracked( plug ) )
+			self.assertEqual( tracker.context( plug ), tracker.context( script["node"] ) )
+
+		# Edit the context. This should also emit the view-specific signal.
+
+		with self.UpdateHandler() :
+			script["contextVariables"]["variables"][0]["value"].setValue( False )
+
+		self.assertEqual( len( genericChangedSlot ), 1 )
+		self.assertEqual( len( specificChangedSlot ), 2 )
+
+		for plug in ( view["in"], view["testPlug"] ) :
+			self.assertFalse( tracker.isTracked( plug ) )
+			self.assertEqual( tracker.context( plug ), tracker.context( script["node"] ) )
+
 if __name__ == "__main__":
 	unittest.main()